refactor: #1057 refactors Label component and removes hover state on remove button
